refactor(core): create adapters and ports folders and move files
