+++ src/app/core/ports/authentication.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { RegisterResponse } from '@core/models/interfaces/register-response.interface';
 import { LoginResponse } from '@core/models/interfaces/login-response.interface';
-import { AuthenticationFirebaseService } from '@core/services/authentication-firebase.service';
+import { AuthenticationFirebaseService } from '@core/adapters/authentication-firebase.service';
 
 @Injectable({
   providedIn: 'root',
