+++ src/app/core/ports/user.service.spec.ts
@@ -1,12 +1,16 @@
 import { TestBed } from '@angular/core/testing';
 
 import { UserService } from './user.service';
+import { provideHttpClient } from '@angular/common/http';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 
 describe('UserService', () => {
   let service: UserService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      providers: [provideHttpClient(), provideHttpClientTesting()],
+    });
     service = TestBed.inject(UserService);
   });
